requirements.txt
Flask
nlpcloud
python-dotenv
Werkzeug
gunicorn